+++ RealEstateBusinessLogicObject/ImageBLO.cs
@@ -62,8 +62,12 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row have just updated</returns>
         public override int Update(RealEstateDataContext.IMAGE entity)
         {
-            _db.Update(entity);
-            return entity.ID;
+            if (ValidationID(entity.ID))
+            {
+                _db.Update(entity);
+                return entity.ID;
+            }
+            else throw new RealEstateDataContext.Utility.ImageID();
         }
 
         /// <summary>
@@ -76,14 +80,18 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row have just updated</returns>
         public int Update(int id, string name, string path, string description)
         {
-            RealEstateDataContext.IMAGE entity = new RealEstateDataContext.IMAGE();
-            entity.ID = id;
-            entity.Name = name;
-            entity.Path = path;
-            entity.Description = description;
+            if (ValidationID(id))
+            {
+                RealEstateDataContext.IMAGE entity = new RealEstateDataContext.IMAGE();
+                entity.ID = id;
+                entity.Name = name;
+                entity.Path = path;
+                entity.Description = description;
 
-            _db.Update(entity);
-            return entity.ID;
+                _db.Update(entity);
+                return entity.ID;
+            }
+            else throw new RealEstateDataContext.Utility.ImageID();
         }
 
         /// <summary>
@@ -93,7 +101,11 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row has just deleted</returns>
         public override void Delete(int ID)
         {
-            _db.Delete(ID);
+            if (ValidationID(ID))
+            {
+                _db.Delete(ID);
+            }
+            else throw new RealEstateDataContext.Utility.ImageID();
         }
 
         /// <summary>
@@ -103,7 +115,11 @@ namespace RealEstateBusinessLogicObject
         /// <returns>Entity</returns>
         public override RealEstateDataContext.IMAGE GetARecord(int ID)
         {
-            return _db.GetARecord(ID);
+            if (ValidationID(ID))
+            {
+                return _db.GetARecord(ID);
+            }
+            else throw new RealEstateDataContext.Utility.ImageID();
         }
     }
 }
