+++ RealEstateBusinessLogicObject/News_TypeBLO.cs
@@ -41,7 +41,7 @@ namespace RealEstateBusinessLogicObject
         /// Insert a row into NEWS_TYPE table
         /// </summary>
         /// <param name="name">News type's name</param>
-        /// <param name="description">News type's </param>
+        /// <param name="description">News type's description</param>
         /// <returns>ID of row has just inserted</returns>
         public int Insert(string name, string description)
         {
@@ -61,8 +61,12 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row has just updated</returns>
         public override int Update(RealEstateDataContext.NEWS_TYPE entity)
         {
-            _db.Update(entity);
-            return entity.ID;
+            if (ValidationID(entity.ID))
+            {
+                _db.Update(entity);
+                return entity.ID;
+            }
+            else throw new RealEstateDataContext.Utility.News_TypeID();
         }
 
         /// <summary>
@@ -74,13 +78,17 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row has just updated</returns>
         public int Update(int id, string name, string description)
         {
-            RealEstateDataContext.NEWS_TYPE entity = new RealEstateDataContext.NEWS_TYPE();
-            entity.ID = id;
-            entity.Name = name;
-            entity.Description = description;
+            if (ValidationID(id))
+            {
+                RealEstateDataContext.NEWS_TYPE entity = new RealEstateDataContext.NEWS_TYPE();
+                entity.ID = id;
+                entity.Name = name;
+                entity.Description = description;
 
-            _db.Update(entity);
-            return entity.ID;
+                _db.Update(entity);
+                return entity.ID;
+            }
+            else throw new RealEstateDataContext.Utility.News_TypeID();
         }
 
         /// <summary>
@@ -90,7 +98,11 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row has just deleted</returns>
         public override void Delete(int ID)
         {
-            _db.Delete(ID);
+            if (ValidationID(ID))
+            {
+                _db.Delete(ID);
+            }
+            else throw new RealEstateDataContext.Utility.News_TypeID();
         }
 
         /// <summary>
@@ -100,7 +112,11 @@ namespace RealEstateBusinessLogicObject
         /// <returns>Entity</returns>
         public override RealEstateDataContext.NEWS_TYPE GetARecord(int ID)
         {
-            return _db.GetARecord(ID);
+            if (ValidationID(ID))
+            {
+                return _db.GetARecord(ID);
+            }
+            else throw new RealEstateDataContext.Utility.News_TypeID();
         }
     }
 }
